Moved multer configuration to a separate file
